+++ mrb_doc/text.rb
@@ -1,12 +1,3 @@
-# Unloads a font from memory
-# @param font [Font]
-# @return [nil]
-def unload_font(font)
-  # mrb_unload_font
-  # src/mruby_integration/text.cpp
-  nil
-end
-
 # Draws the current frame rate at the passed in x and y coordinates.
 # @param x [Integer]
 # @param y [Integer]